chore(libsy): move RoutedLlmClient and Decision traits to protocol crate
