+++ crates/libsy-protocol/Cargo.toml
@@ -17,6 +17,7 @@ keywords = ["llm", "protocol", "openai", "anthropic"]
 publish = ["crates-io"]
 
 [dependencies]
+async-trait = "0.1"
 futures = "0.3"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
