+++ crates/libsy-protocol/src/lib.rs
@@ -20,11 +20,13 @@
 //!
 //! [`Algorithm`]: https://docs.rs/libsy
 
+pub mod client;
 pub mod envelope;
 pub mod format;
 pub mod llm;
 pub mod stream;
 
+pub use client::*;
 pub use envelope::*;
 pub use format::*;
 pub use llm::*;
